ingestion/requirements.txt
ingestion
requests==2.26.0
bs4==0.0.1
openpyxl==3.0.9
pandera==0.8.0
SQLAlchemy==1.4.28
pyyaml==6.0